+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 0.1.2
 -----
 - added forced re-login for clicking an active account after N hours.
-
+- implemented search functionality with aliases
 0.1.1
 -----
 - light icons for darkmode browsers